examples/servers/time/function/requirements.txt
mcp-server-time

# For testing, this module is installed from local files
# mcp-lambda
